Move start_daemon to daemon::embedded
